Admin and Instructor route groups added along with skeleton admin page as an example placeholder. This will allow us to organize routes to use certain layouts/errors depending upon the user role.
