+++ app/(admin)/adminDash/page.tsx
@@ -1,0 +1,11 @@
+// pages/hello.tsx
+
+const AdminPage: React.FC = () => {
+  return (
+    <div>
+      <h1>This is the admin page!</h1>
+    </div>
+  );
+};
+
+export default AdminPage;
