+++ app/(instructor)/dashboard/class-settings/appearance/components/appearance-form.tsx
