+++ app/(instructor)/dashboard/class-settings/appearance/page.tsx
