+++ app/(instructor)/dashboard/class-settings/layout.tsx
