+++ app/(instructor)/dashboard/class-settings/page.tsx
@@ -1,5 +1,5 @@
 import { Separator } from '@/components/ui/separator';
-import ProfileForm from '@/app/dashboard/class-settings/components/profile-form';
+import ProfileForm from '@/app/(instructor)/dashboard/class-settings/components/profile-form';
 
 export default function SettingsProfilePage() {
   return (
