+++ app/(instructor)/dashboard/components/main-nav.tsx
