+++ app/(instructor)/dashboard/manage-course-members/columns.tsx
