+++ app/(instructor)/dashboard/take-attendance/components/data-table-toolbar.tsx
@@ -5,7 +5,7 @@ import { Table } from '@tanstack/react-table';
 
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import { DataTableViewOptions } from '@/app/dashboard/take-attendance/components/data-table-view-options';
+import { DataTableViewOptions } from '@/app/(instructor)/dashboard/take-attendance/components/data-table-view-options';
 
 import { statuses } from '../data/data';
 import { DataTableFacetedFilter } from './data-table-faceted-filter';
