+++ app/(instructor)/dashboard/testing-playground/components/ErrorTest.tsx
@@ -1,5 +1,4 @@
 'use client';
-import React from 'react';
 import { Button } from '@/components/ui/button';
 
 const ErrorTestingPage = () => {
