+++ app/(instructor)/qr/DynamicQRCodeComponent.tsx
